X11::createWmWindow  int event_mask modified to ulong event_mask

+++ src/X11.hh
@@ -447,7 +447,7 @@ public:
 					 ulong border, ulong background);
 	static Window createWmWindow(Window parent,
 				     int x, int y, uint width, uint height,
-				     int event_mask);
+				     ulong event_mask);
 	static void destroyWindow(Window win);
 	static void changeWindowAttributes(Window win, ulong mask,
 					   XSetWindowAttributes &attrs);
